+++ GenericQuest/Actor.h
@@ -14,7 +14,7 @@ class Actor
 		virtual void update(float delta);
 		virtual void draw(Canvas* canvas);
 		virtual void setDimension(int x, int y);
-		virtual void setPosition(int x, int y);
+		virtual void setPosition(float x, float y);
 		virtual void setVelocity(float x, float y);
 		virtual void setAcceleration(float x, float y);
 		virtual void setHidden(bool newHidden);
