+++ GenericQuest/Canvas.h
@@ -9,7 +9,7 @@ class Canvas
 	public:
 		Canvas();
 		void overwrite(Frame* frame);
-		void overwrite(Frame* frame, int xPos, int yPos);
+		void overwrite(Frame* frame, float xPos, float yPos);
 		void draw(HANDLE screen);
 		void clear();
