+++ GenericQuest/Frame.h
@@ -9,8 +9,8 @@ class Frame : public Actor
     public:
 		Frame();
 		Frame(int width, int height);
-		Frame(string file, int x, int y);
-		Frame(vector<vector<Symbol>> newFrame, int x, int y);
+		Frame(string file, float x, float y);
+		Frame(vector<vector<Symbol>> newFrame, float x, float y);
 		~Frame();
 		virtual void update(float delta);
 		virtual void draw(Canvas* canvas);
@@ -19,7 +19,7 @@ class Frame : public Actor
 		virtual void setBackgroundColor(int color);
 		virtual void setSymbol(Symbol symbol, int x, int y);
 		virtual void setDimension(int x, int y);
-		virtual void setPosition(int x, int y);
+		virtual void setPosition(float x, float y);
 		virtual void setVelocity(float x, float y);
 		virtual void setAcceleration(float x, float y);
 		virtual Symbol getSymbol(int x, int y);
