+++ GenericQuest/Menu.h
@@ -7,7 +7,7 @@
 class Menu : public Actor
 {
 	public:
-		Menu(Frame* newMessage, Frame* newCursor, Frame* member, int x, int y, char newUp, char newDown);
+		Menu(Frame* newMessage, Frame* newCursor, Frame* member, float x, float y, char newUp, char newDown);
 		~Menu();
 		virtual void update(float delta);
 		virtual int input();
@@ -16,7 +16,7 @@ class Menu : public Actor
 		virtual void setForegroundColor(int color);
 		virtual void setBackgroundColor(int color);
 		virtual void setDimension(int x, int y);
-		virtual void setPosition(int x, int y);
+		virtual void setPosition(float x, float y);
 		virtual void setVelocity(float x, float y);
 		virtual void setAcceleration(float x, float y);
 		virtual Dimension getDimension();
@@ -28,7 +28,7 @@ class Menu : public Actor
 		vector<Frame*> members;
 		Frame* message;
 		Frame* cursor;
-		int focusedMember;
+		unsigned int focusedMember;
 		char up, down;
 };
