+++ GenericQuest/Text.h
@@ -9,13 +9,13 @@ class Frame;
 class Text : public Frame
 {
 	public:
-		Text(bool useFile, string file, bool type, int newSpeed, int newPause, int x, int y);
+		Text(bool useFile, string file, bool type, int newSpeed, int newPause, float x, float y);
 		~Text();
 		virtual void update(float delta);
 		virtual void draw(Canvas* canvas);
 		virtual void setSymbol(Symbol symbol, int x, int y);
 		virtual void setDimension(int x, int y);
-		virtual void setPosition(int x, int y);
+		virtual void setPosition(float x, float y);
 		virtual void setVelocity(float x, float y);
 		virtual void setAcceleration(float x, float y);
 		virtual Symbol getSymbol(int x, int y);
@@ -27,7 +27,7 @@ class Text : public Frame
 	private:
 		vector<vector<Symbol>> text; //Holds the text
 		Timer timer;
-		int typeSpeed, pause, cursorX, cursorY;
+		unsigned int typeSpeed, pause, cursorX, cursorY;
 		bool typewriter, paused;
 };
